Add files via upload
This file includes Data visualization using Matplotlib and Pandas.The dataset used here is the Company employee details. The Visualization has been done between the columns: Years_in _the _company and age_when_joined. This visualization will reflect the relationship between the years that the employees have worked in the copany so far and the age at the time of their joining the company. Thank you. Regards Smriti